Change pageState.pageOutputs type
Use a map[output.Format]*pageOutput rather than a []*pageOutput.
This is to make associations between Site.renderFormats and
pageState.outputFormats clearer. In *HugoSites.render, for example,
we find a pageOutput that matches a render format via
a shared index, which is difficult to understand if you don't
already know that pageOutputs has the same length and order
renderFormats.