Return to previous state on config reload error
Previously, fullRebuild created a newCommandeerHugoState and
assigned that to c.commandeerHugoState, then attempted to load
the config. If that failed, it would leave c.commandeerHugoState
as the new commandeerHugoState, even though loadConfig is what
populates this--i.e., c.hugoSites would be nil. The watcher
reilies on c.hugoSites to be non-nil, so this would break the
file system watcher.

Since we need to successfully load the config to rebuild the
HugoSites and watch for changes, this commit reverts to
the previous hugo state if the config loading fails.